Interrupts + DF handler

+++ src/types.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 typedef unsigned long long uint64_t;
-typedef unsigned long int uint32_t;
+typedef unsigned int uint32_t;
 typedef unsigned short uint16_t;
 typedef unsigned char uint8_t;
